Adding table search framework and use in admin instances
This adds table seaching on server side as a new FilterAction type.
It also implements table search by project on the instances panel in
admin dashboard as a sample use of the framework. This will help
admins when wading through the list of all instances in the cloud.

Implements: blueprint table-filtering

Change-Id: I337dccb51cbd0b1a477606ac97b17dea3295960f