Restrict flavor fields input on creation/update.
If flavor properties are invalid, an update message error
is shown, but the flavor get deleted.
The reason is that update flavor is not yet supported so horizon
will delete the flavor and create a new one, failing in the last case.

Fields ram and vcpus need to be >= 1 and eph_gb, swap_mb and disk_gb
>= 0 to create a flavor.

Change-Id: I20d178bad6e4e50c0963acdec72ebedb9bbaba17
Closes-Bug: #1269638