Implementing accordion navigation
Left hand side navigation is now an accordion with collapsable
dashboards and panel groups.

The project and region pickers have been relocated to the header.
They will be further enhanced in a future bp.

The help link and settings link have now been placed in a user
dropdown as well.

The page header text is still preserved and now sits above the
main content. This information is generally redundant with the
table headers or the panel selected in the left hand navigation,
and should be considered for removal in a different bp.

Implements blueprint navigation-enhancement

Change-Id: I9588cabdf4ce41bd08cadff780b1a87f73dbbea5