Policy check needs to check for None in target
For objects that don't have a project_id attached to them, the
property can be set to None, which does not get overridden in
the policy engine where it being not set currently does.  This
adds a check for user_id and project_id being None and setting them
if they are.

Change-Id: I7aeb6d3830a19a7191de9944f8de90ee12dbf127
Closes-bug: #1274053