Fixes instance double delete via row_action
Currently when an instance is terminated, via the project/instances view,
a "Terminate Instnace" button appears in the row_action column.

This is due to a the TerminateInstance class's allowed() method always
returning true.

By putting logic in this method to only allow the terminate row_action
when an instance is *not* being deleted, we prevent the double delete from
occurring via row_action.

Change-Id: I23e759036c8f6c35331243717853c8a5e269fa75
Closes-Bug: 1278487